+++ lib/stdio.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
-int sprintf(char *s, const char *fmt, ...) {
+long sprintf(char *s, const char *fmt, ...) {
     va_list args;
-    int i = 0;
+    long i = 0;
 
     va_start(args, fmt);
     i = vsprintf(s, fmt, args);
