+++ platform/virt/mm.c
@@ -1,5 +1,5 @@
 #include <kernel/mm.h>
 
-unsigned int PAGESIZE = 0x1000;
-unsigned int MEMBASEADDR = 0x40000000;
-unsigned int MEMSIZE = 0x20000000;
+unsigned long PAGESIZE = 0x1000;
+unsigned long MEMBASEADDR = 0x40000000;
+unsigned long MEMSIZE = 0x20000000;
